Move the workspace layout reducer tests to their conventional file
